Extract reserved identifier renaming to separate module (476 lines)
