+++ src/plugin/src/transforms/feather/index.ts
@@ -4,3 +4,4 @@
 export * from "./ast-traversal.js";
 export * from "./enum-handling.js";
 export * from "./parser-bootstrap.js";
+export * from "./reserved-identifier-renaming.js";
